requirements.txt
streamlit
sqlalchemy
bcrypt
spacy
pymupdf
python-docx
fpdf2
scikit-learn
google-genai